requirements.txt
Flask==2.0.1
Pillow==9.5.0
